+++ src/commands/markdown-commands/headingLevel1Command.ts
@@ -1,9 +1,8 @@
-import * as React from 'react';
 import { Command } from '../command';
 import { setHeader } from '../../helpers/headerHelpers';
 
-export const headingLevel6Command: Command = {
+export const headingLevel1Command: Command = {
   execute: ({ initialState, textApi }) => {
-    setHeader(initialState, textApi, '###### ');
+    setHeader(initialState, textApi, '# ');
   },
 };
