+++ src/commands/markdown-commands/headingLevel2Command.ts
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import { Command } from '../command';
 import { setHeader } from '../../helpers/headerHelpers';
