--- src/commands/markdown-commands/list.tsx
@@ -1,1 +1,0 @@
-import * as React from 'react';
