+++ src/commands/markdown-commands/quoteCommand.ts
@@ -1,4 +1,3 @@
-import * as React from 'react';
 import { Command } from '../command';
 import {
   getBreaksNeededForEmptyLineAfter,
@@ -7,8 +6,8 @@ import {
   selectWord,
 } from '../../helpers/textHelpers';
 
-export const codeBlockCommand: Command = {
-  execute: async ({ textApi, initialState }) => {
+export const quoteCommand: Command = {
+  execute: ({ initialState, textApi }) => {
     // Adjust the selection to encompass the whole word if the caret is inside one
     const newSelectionRange = selectWord({
       text: initialState.text,
@@ -16,30 +15,16 @@ export const codeBlockCommand: Command = {
     });
     const state1 = textApi.setSelectionRange(newSelectionRange);
 
-    // when there's no breaking line
-    if (getSelectedText(state1).indexOf('\n') === -1) {
-      textApi.replaceSelection(`\`${getSelectedText(state1)}\``);
-      // Adjust the selection to not contain the **
-
-      const selectionStart = state1.selection.start + 1;
-      const selectionEnd = selectionStart + getSelectedText(state1).length;
-
-      textApi.setSelectionRange({
-        start: selectionStart,
-        end: selectionEnd,
-      });
-      return;
-    }
-
     const breaksBeforeCount = getBreaksNeededForEmptyLineBefore(state1.text, state1.selection.start);
     const breaksBefore = Array(breaksBeforeCount + 1).join('\n');
 
     const breaksAfterCount = getBreaksNeededForEmptyLineAfter(state1.text, state1.selection.end);
     const breaksAfter = Array(breaksAfterCount + 1).join('\n');
 
-    textApi.replaceSelection(`${breaksBefore}\`\`\`\n${getSelectedText(state1)}\n\`\`\`${breaksAfter}`);
+    // Replaces the current selection with the quote mark up
+    textApi.replaceSelection(`${breaksBefore}> ${getSelectedText(state1)}${breaksAfter}`);
 
-    const selectionStart = state1.selection.start + breaksBeforeCount + 4;
+    const selectionStart = state1.selection.start + breaksBeforeCount + 2;
     const selectionEnd = selectionStart + getSelectedText(state1).length;
 
     textApi.setSelectionRange({
