+++ src/commands/markdown-commands/strikethroughCommand.ts
@@ -1,8 +1,7 @@
-import * as React from 'react';
 import { Command } from '../command';
 import { getSelectedText, selectWord } from '../../helpers/textHelpers';
 
-export const linkCommand: Command = {
+export const strikethroughCommand: Command = {
   execute: ({ initialState, textApi }) => {
     // Adjust the selection to encompass the whole word if the caret is inside one
     const newSelectionRange = selectWord({
@@ -10,12 +9,12 @@ export const linkCommand: Command = {
       selection: initialState.selection,
     });
     const state1 = textApi.setSelectionRange(newSelectionRange);
-    // Replaces the current selection with the bold mark up
-    const state2 = textApi.replaceSelection(`[${getSelectedText(state1)}](url)`);
-    // Adjust the selection to not contain the **
+    // Replaces the current selection with the strikethrough mark up
+    const state2 = textApi.replaceSelection(`~~${getSelectedText(state1)}~~`);
+    // Adjust the selection to not contain the ~~
     textApi.setSelectionRange({
-      start: state2.selection.end - 6 - getSelectedText(state1).length,
-      end: state2.selection.end - 6,
+      start: state2.selection.end - 2 - getSelectedText(state1).length,
+      end: state2.selection.end - 2,
     });
   },
 };
