+++ src/types/CommandOptions.ts
@@ -1,34 +1,27 @@
-import { SelectionRange } from './SelectionRange';
+export interface SelectionRange {
+  start: number;
+  end: number;
+}
 
-/**
- * The state of the text of the whole editor
- */
+// The state of the text of the whole editor
 export interface TextState {
-  /**
-   * All the text in the editor
-   */
+  // All the text in the editor
   text: string;
-  /**
-   * The section of the text that is selected
-   */
+
+  // The section of the text that is selected
   selection: SelectionRange;
 }
 
 export interface TextController {
-  /**
-   * Replaces the current selection with the new text. This will make the new selectedText to be empty, the
-   * selection start and selection end will be the same and will both point to the end
-   * @param text Text that should replace the current selection
-   */
+  // Replaces the current selection with the new text. This will make the new selectedText to be empty, the
+  // selection start and selection end will be the same and will both point to the end
+  // @param text Text that should replace the current selection
   replaceSelection(text: string): TextState;
 
-  /**
-   * Selects the specified text range
-   * @param selection
-   */
+  // Selects the specified text range
+  // @param selection
   setSelectionRange(selection: SelectionRange): TextState;
-  /**
-   * Get the current text state
-   */
+
+  // Get the current text state
   getState(): TextState;
 }
