--- src/types/SelectionRange.ts
@@ -1,4 +1,0 @@
-export interface SelectionRange {
-  start: number;
-  end: number;
-}
